[BACKEND] Removed ValueRef thunks.
Previously, 'execucuteXXX' methods from the runtime API
returned a DataBag wrapped in a ValueRef. This however, caused problems
when passing closure parameters to the dataflow.

With this rewrite, the runtime API now returns 'DataBag' instances
which can be cast to 'ParallelizedDataBag' in order to obtain information
about their parallel backing representations (Flink RDD or Spark DataSet).