[MACROS] Fixed bugs in the ComprehensionCombination.
The current rewrite system has two bugs.

First, the guards of the `Join` and `Cross` rules should ensure
that the two matched combinators are not correlated.

Second, it does not accomodate for residual comprehensions with more
than one correlated generators upon leaving the `Cross` state.

To fix the first issue the code was changed as follows:

1. Generalize the `UnnestHead` rule in the normalization engine in
order to cover arbitrary head types. This ensures that all `MonadJoin`
nodes are removed during the normalization phase.

2. Rework the `MatchFlatMap` rule to work on pairs of generators, thus
accomodating for tha changes of UnnestHead in (1).

3. Pull the `MatchFlatMap` rule in a dedicated phase in the rewrite
engine state machine right after `Filter`. This ensures that correlated
generators will be combined as `flatMap` with precedence.

4. Implement `guard` logic for `Join` and `Cross` combinators which
prohibits rule application on correlated inputs.

5. Rework the final phase so it accomodates for residual comprehensions
with more than one generator.