[MACROS] Replace substitution with symbol binding
We step on the equivalence of the following two constructs (shown
in untyped lambda calculus for simplicity):

    expr[t\x] = (\x. expr) t

I.e. substitution and binding followed by subsequent application
are beta-equivalent. Instead of beta-reduction, we can use
abstraction, which is cleaner to program in macros.

In Scala the right-hand side can be simplified to a local variable
definition inside a block instead of an anonymous function (with
careful renaming in case `x` is free in `t`:

    { val fresh = t; expr[fresh\x] }

The advantage is that renaming can safely be performed on symbols.
Furthermore, the program execution is potentially more efficient
and follows closely the original (`t` is executed only once).

Open issue: handle `substitute(find: Tree, repl: Tree)`