[MACROS] Made macro-level IR immutable.
Resolves #153.

Adapted the ComprehensionModel nodes so all fields are immutable.

The only place where mutability is left is in the Generator type, as
otherwise we would have to also rewrite the FoldGroupFusion optimization,
which won't be trivial and at the moment seems unnecessary.
The Generator type should be fixed as we make further progress on #147.