[MACROS] Migrated runtime.universe to JavaUniverse
The refined type allows access to tree mutating methods, such as `setType`, `setSymbol`, `setOwner`, `setInfo`, `setOriginal` and `attachments`, that are otherwise hidden from the APIs in trait `Universe`. Since the project is already committed to the JVM, this poses no runtime limitations.

Note: `RuntimeCompiler` had to be modified to create a `ToolBox` of the expected type.