[MACROS] Added missing mutators to compiler.Util
Some mutator methods for `Tree`s and `Symbol`s are hidden in the top `Universe` trait. Those are added as abstract wrappers to `compiler.Util` to be overridden by subtraits specialized for a refined universe type. However, they are package private in `compiler` since they are to be used only as helper methods and should not be part of the API.