[IR] Reworked alpha-equivalence for Scala ASTs.
The new version resides under `compiler.lang.AlphaEq.alphaEq`
and uses `scalactic` for shortcut semantics. The return type
is either `True`, or an `Error` which contains a `msg` with
a description of the first problem and the `lhs` and `rhs`
trees where it occurred.

Trees should be from now on compared with

```
lhs shouldBe alphaEqTo(rhs)
```

instead of the old

```
lhs shouldEqual
```

In addition, this also introduces the following changes.

- Introduced a custom `BeMatcher` called `alphaEqTo`.
- Removed the current `TreeEquality` traits.
- Removed the current `eq` implementations.