[IR] Eliminated temp variables from `dscf`.
The `dscf` transformation currently unnecessarily binds control-flow
expressions which can be directly included in the expr position of the
resulting blocks to synthetic valdefs which are referenced in the block expr.

Due to this, the result is not a valid LNF tree and needs to go through an
extra `inlineLetExprs` transformation in order to become valid LNF' term.

The modifications in this commit resolve this issue.