[API] DataBag API revision.
Declarted `DataBag` as an abstract trait and added a generic `DataBagSpec`
that is to be used by implementations.

Implementers of the `DataBag` trait can be deep or shallow embeddings that
abstracting over different backends (e.g. Spark's RDD or Dataset APIs or
Flink's DataSet API).

The first stab includes a reference `DataBag` implementation backed by a
Scala `Traversable` representation that can be used for local processing.

This revision of the API, as compared to the old version which resides
under `eu.stratosphere.emma.api.DataBag` includes the following changes.

- Renamed `plus` to `union`.
- Removed `minus` (not supported by most targeted backends).