[API] Made specific folds an explicit part of the new `DataBag` API.
This basically inlines the `FoldMacros` from the old API directly within
the `DataBag` API. I think that it might be beneficial to hold on to the
combinator form for a while and not bluntly substitute before entering
the compiler pipeline.

The resolution of various FIXME comments in the "structural recursion"
spec was enabled by the migration to non-macro fold definitions.