[API] Simplified `org.emmalanguage.api` package object.
1. Moved `Meta` and `CSVConverter` implicits to companion objects.

Based on a great hint by @joroKr21, this gem of a refactoring allows
us to port the new `DataBag` API to the new compiler infrastructure
with zero modifications in the `BaseCodegenTest.checkCompile` method.

Implicit type instance resolution now kicks in in the second phase,
as described in this Stackoverflow post:

http://stackoverflow.com/q/5598085

2. Renamed the implicit `ttag` and `ctag` projections.

Renamed as `ttagFor` and `ctagFor`.

3. Removed obsolete `Limits` trait.

These were used initially for the min / max, but are no longer needed.