[IR] Fixed type assertions `DefDef.apply`.
Type assertions in `DefDef.apply` are failing in the presence of type
parameters. For example, performing an identity transformation that
simply matches and re-creates DefDef nodes on a snippet such as

```
def plus[A : Numeric](x: A, y: A): A =
  implicitly[Numeric[A]].plus(x, y)
```

would fail due to type errors in `x`, `y`, and in the `DefDef` body.

To fix the assertion behavior, types are now be replaced before the checks.