[SPARK][FLINK] Factored out backend-specific compiler logic.
Factored out backend-specific compiler logic in dedicated traits
called `SparkCompiler` and `FlinkCompiler`.

Those can be mixed in in specialized macros (e.g., `SparkMacro`
and `FlinkMacro`), as well as in specialized tests (such as the
subclasses of `BaseCodegenIntegrationSpec`).

Backend-specific compiler traits install extra reflection modules
and extend the `implicitTypes` field as follows:

- `FlinkCompiler` adds `ExecutionEnvironment` and `TypeInformation`;
- `SparkCompiler` adds `SparkSession` and `Encoder`.

This allows to prune the implicit parameter lists from the `SparkRDD`
and `SparkDataset` constructor calls.