[API][FLINK][SPARK][IR] Added `MutableBag` abstraction.
This is a somewhat simplified port of the `StatefulBag` abstraction
from the old master branch.

The implementation is based on an abstract trait (`MutableBag`) with
concrete implementations for Flink and Spark.

A `MutableBag` can be constructed from a set of key-value pairs of
type `(K, V)`.

Subsequent bulk-updates on a mutable bag `xs` are possible with a
bag of messages `ms` of type `DataBag[Group[K, M]]` and an update
function `f` of type `(K, Option[V], M) => Option[V]`.

For each `m` in `ms`, the `update` implementation will call

- `f(m.key, Some(v), m.value)` iff `(m.key, v)` is in `xs`;
- `f(m.key, None, m.value)`    iff `(m.key, v)` is not in `xs`.

Each `f` call can optionally return an `v` value to be associated
with the current `k`.

The collection of updated values is merged in the underlying mutable
bag and returned as final result.

The functional semantics of the function rely on the assumption that
the first component (of type `K`) is a unique key in `ms`. Otherwise,
the implied state update is non-deterministic.

Assigning of `MutableBag` instances has by-reference semantics and
subsequent `update` calls will also affect the original instance. To
enforce `by-value` assignment, use the `val ys = xs.copy()`.