[IR] Fixed some bugs in the `comprehension` package.
The `resugar, the `lookup` case for externally defined lambdas
`f: A => B` has a wront argument type `B` for `x` (should be `A`).

In addition, the rhs of the returned pair is not valid Emma Core
expression, which propagades to the resugared guards. To fix this,
instead of `f(x)`, the rhs is modified as `{ val r = f(x); r}`.

The same problem occcurs in the `AsComprehension` cast during
comprehension normalization, where the rhs of the comprehension
should be `{ val r = tree; r }` instead of just `{ tree }`.

To handle both of these issues, the `asLet` method now produces
valid Emma Core blocks.