[IR,API] Added ad-hoc compiler pipeline configuration support.
At the moment, the compiler pipeline defined by the `emma.onSpark`
and `emma.onFlink` macros is fixed.

This commit adds overloaded variants of these methods which take
an additional `config` parameter. This can be used to point to a
HOCON config which overrides certain compiler pipeline parameters,
e.g. application of optimizations such as `foldFusion`.