[IR] Fixed wrong highlighting of path-dependent `u` types in IDEA.
Most notably, this was caused by type definitions placed within an
object nested in an enclosing compiler trait.

For example:

- `Syntax` declared in the `Comprehension` object.
- `CG` and `LibDefRegistry` declared in the `LibSupport` object.
- `FlowGraph` declared in `CFG`.

Moving those declarations direcly under the enclosing trait solved
the highlighting issues.

In addition, the PR unifies the self-type name (`this` => `self`)
and re-organizes the imports in the `compiler` package.