[IR] Reworked comprehension normalization (1/3).
Changed resugaring of `xs.flatMap(x => f(x))` to produce

```scala
for (x1 <- xs; x2 <- { f(x) }) yield { x2 }
```

instead of

```scala
flatten(for (x <- xs) yield f(x))
```

This alleviates the need of a dedicated `flatten` operator
in the mock-comprehension syntax.

To accomodate for the change in the `Normalize` module, also
removed `UnnestHead` (which is now done implicitly in the
resugar step), and added a second variant of `UnnestHead`.