[IR] Fix a bug in `FoldGroupFusion`.
The rewrite should not pull `m.d` values before the infix
of values occurring between `x` and `c`, as some of the
values in `m.d` might depend on those.

Instead, the definition of `x` should move just before the
definition of `c`. This is sound, as we know that `x` is
used only once.